+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- Browser auto-reload is now the default behavior as that's what I personally find most useful. With this the `-w`/`--watch` option has been removed. A new `--no-auto-reload` option has been added to disable auto-reloading for cases where people prefer to manually refresh their browsers to see updates.
+- Automatic browser refresh is now the default behavior as that's what I personally find most useful. With this the `-w`/`--watch` option has been removed. A new `--no-auto-refresh` option has been added to disable auto-refreshing for cases where people prefer to manually refresh their browsers to see updates.
 
 ## 21.3.0 - 2021-03-28